+++ components/skus/browser/BUILD.gn
@@ -17,17 +17,17 @@ static_library("browser") {
     "pref_names.h",
     "skus_sdk_impl.cc",
     "skus_sdk_impl.h",
-    "br-rs/brave-rewards-cxx/src/cxx.h",
-    "br-rs/brave-rewards-cxx/src/shim.hpp",
-    "br-rs/brave-rewards-cxx/src/wrapper.cpp",
-    "br-rs/brave-rewards-cxx/src/wrapper.hpp",
+    "brave-rewards-cxx/src/cxx.h",
+    "brave-rewards-cxx/src/shim.h",
+    "brave-rewards-cxx/src/wrapper.cc",
+    "brave-rewards-cxx/src/wrapper.h",
   ]
 
   # TODO(bsclifton): troubleshoot why there are duplicate symbols on Windows
   if (is_mac || is_linux) {
-    sources += [ "br-rs/brave-rewards-cxx/src/cxx.cc" ]
+    sources += [ "brave-rewards-cxx/src/cxx.cc" ]
   } else if (is_win) {
-    sources += [ "br-rs/brave-rewards-cxx/src/cxx_win.cc" ]
+    sources += [ "brave-rewards-cxx/src/cxx_win.cc" ]
   }
 
   deps = [
