+++ components/skus/browser/brave-rewards-cxx/Makefile
@@ -1,13 +1,13 @@
 all: examples/out
 
-src/wrapper.hpp: src/lib.rs
-	cxxbridge src/lib.rs --header > src/wrapper.hpp
+src/wrapper.h: src/lib.rs
+	cxxbridge src/lib.rs --header > src/wrapper.h
 
-src/wrapper.cpp: src/lib.rs
-	cxxbridge src/lib.rs > src/wrapper.cpp
+src/wrapper.cc: src/lib.rs
+	cxxbridge src/lib.rs > src/wrapper.cc
 
 examples/wrapper.o: src/wrapper.cpp src/wrapper.hpp
-	g++ $(CFLAGS) -std=c++11 src/wrapper.cpp -I src/ -c  -o examples/wrapper.o
+	g++ $(CFLAGS) -std=c++11 src/wrapper.cc -I src/ -c  -o examples/wrapper.o
 
 ../target/debug/libbrave_rewards_cxx.a: src/lib.rs src/httpclient.rs src/storage.rs
 	cargo build
