+++ components/skus/browser/brave-rewards-cxx/src/cxx_win.cc
