+++ components/skus/browser/brave-rewards-cxx/src/errors.rs
