+++ components/skus/browser/brave-rewards-cxx/src/shim.h
@@ -10,11 +10,14 @@ enum class RewardsResult : uint8_t;
 struct HttpRequest;
 struct HttpResponse;
 struct HttpRoundtripContext;
+class SkusSdkImpl;
+class SkusSdkFetcher;
 
 class RefreshOrderCallbackState {
  public:
   skus::mojom::SkusSdk::RefreshOrderCallback cb;
-  // TODO: possibly store Prefs object here
+  SkusSdkImpl* instance;
+  SkusSdkFetcher* fetcher;
 };
 
 using RefreshOrderCallback = void (*)(RefreshOrderCallbackState* callback_state,
@@ -22,7 +25,7 @@ using RefreshOrderCallback = void (*)(RefreshOrderCallbackState* callback_state,
                                       rust::cxxbridge1::Str order);
 void shim_purge();
 void shim_set(rust::cxxbridge1::Str key, rust::cxxbridge1::Str value);
-const std::string& shim_get(rust::cxxbridge1::Str key);
+::rust::String shim_get(rust::cxxbridge1::Str key);
 
 void shim_scheduleWakeup(::std::uint64_t delay_ms,
                          rust::cxxbridge1::Fn<void()> done);
