+++ components/skus/browser/brave-rewards-cxx/src/storage.rs
@@ -1,5 +1,6 @@
 use crate::{ffi, NativeClient, LOCAL_POOL};
 use brave_rewards::{errors, KVClient, KVStore};
+use tracing::debug;
 
 impl KVClient<NativeClient> for NativeClient {
     fn get_store() -> Result<Self, errors::InternalError> {
@@ -18,10 +19,14 @@ impl KVStore for NativeClient {
         Ok(())
     }
     fn get(&mut self, key: &str) -> Result<Option<String>, errors::InternalError> {
-        let ret = ffi::shim_get(key).to_string_lossy();
+        let ret = ffi::shim_get(key);
+        debug!("KVStore->get finished call to ffi:shim_get");
+        debug!("{:?}", ret);
         Ok(if ret.len() > 0 {
+            debug!("KVStore->get returning value");
             Some(ret.to_string())
         } else {
+            debug!("KVStore->get returning None");
             None
         })
     }
